agent/requirements.txt
langgraph>=0.2.0
langchain-openai>=0.2.0
pymongo>=4.6.0
python-dotenv>=1.0.0
APScheduler>=3.10.0
pytest>=8.0.0
